Restructure blog template files
